26.07.29 작업내역
1. 작업내역
   - ch01 유적 해결사 뮤리스의 스킬 데이터 제작성 완료
	so 변경에 따른 데이터 규격 변경

   - SkillBase에 ApplyCoinEffects() 함수 작성중
	SO 데이터를 읽어서 누구에게 어떤 기능을 적용시킬지 결정하는 함수
	구조체 SkillUseData의 데이터를 받아와서

   - AddKeyword() 와 AddEffect() 함수 통합
	기존의 SO 타입에 따른 진입 함수 분할 방식에서 AddEffect() 하나로 통합하고
	내부에서 SO 타입에 따라서 로직을 호출하는 방식으로 교체함

   - SkillDescriptionBuilder 의 GetActionText() 로직 일부 변경
	데이터가 List에서 단일 데이터로 바뀌어 데이터 작성 방식 변경

   - ch01 유적 해결사 뮤리스의 스킬 강화 방식 정립
	스킬 강화의 경우, 전투 시작 시 Character_RuinsFixerMurris가
	자신의 가장 왼쪽 스킬 슬롯부터 스킬을 교체하는 방식으로 동작 목표

2. 필요한 작업
   - 스킬 강화 로직 구현 필요
   - 오리지널 이펙트 동작 함수 구현 필요

+++ Assets/Scripts/Character/Skill/SkillSO.cs
@@ -68,7 +68,7 @@ public class SkillSO : ScriptableObject
         /// <summary>
         /// 해당 스킬의 데미지 값
         /// </summary>
-        public int motionValue;
+        public float motionValue;
         /// <summary>
         /// 스킬의 기본 위력
         /// </summary>
